requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
aiosignal>=1.3.1
attrs>=22.2.0
